requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
groq==0.4.2
werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
